+++ libs/base/src/text/char/tests.rs
@@ -1,4 +1,4 @@
-// devela::text::char::tests
+// devela_base::text::char::tests
 //
 //! These fns tests all the character types
 //