+++ src/text/grapheme/string_u8.rs
@@ -30,47 +30,47 @@ impl<const CAP: usize> GraphemeU8<CAP> {
         Ok(Self(unwrap![ok? StringU8::new()]))
     }
 
-    // /// Creates a new `GraphemeU8` from a `char7`.
-    // ///
-    // /// # Errors
-    // /// Returns [`MismatchedCapacity`] if `CAP > 255.
-    // ///
-    // /// Will always succeed if `CAP` >= 1 and <= 255.
-    // pub const fn from_char7(c: char7) -> Result<Self, MismatchedCapacity> {
-    //     Ok(Self(unwrap![ok? StringU8::from_char7(c)]))
-    // }
-    //
-    // /// Creates a new `GraphemeU8` from a `char8`.
-    // ///
-    // /// # Errors
-    // /// Returns [`MismatchedCapacity`] if `CAP` > 255
-    // /// or < `c.`[`len_utf8()`][char8#method.len_utf8].
-    // ///
-    // /// Will always succeed if `CAP` >= 2 and <= 255.
-    // pub const fn from_char8(c: char8) -> Result<Self, MismatchedCapacity> {
-    //     Ok(Self(unwrap![ok? StringU8::from_char8(c)]))
-    // }
-    //
-    // /// Creates a new `GraphemeU8` from a `char16`.
-    // ///
-    // /// # Errors
-    // /// Returns [`MismatchedCapacity`] if `CAP` > 255
-    // /// or < `c.`[`len_utf8()`][char16#method.len_utf8].
-    // ///
-    // /// Will always succeed if `CAP` >= 3 and <= 255.
-    // pub const fn from_char16(c: char16) -> Result<Self, MismatchedCapacity> {
-    //     Ok(Self(unwrap![ok? StringU8::from_char16(c)]))
-    // }
-    //
-    // /// Creates a new `GraphemeU8` from a `char`.
-    // /// # Errors
-    // /// Returns [`MismatchedCapacity`] if `CAP` > 255
-    // /// or < `c.`[`len_utf8()`][UnicodeScalar#method.len_utf8].
-    // ///
-    // /// Will never panic if `CAP` >= 4 and <= 255.
-    // pub const fn from_char(c: char) -> Result<Self, MismatchedCapacity> {
-    //     Ok(Self(unwrap![ok? StringU8::from_char(c)]))
-    // }
+    /// Creates a new `GraphemeU8` from a `char7`.
+    ///
+    /// # Errors
+    /// Returns [`MismatchedCapacity`] if `CAP > 255.
+    ///
+    /// Will always succeed if `CAP` >= 1 and <= 255.
+    pub const fn from_char7(c: char7) -> Result<Self, MismatchedCapacity> {
+        Ok(Self(unwrap![ok? StringU8::from_char7(c)]))
+    }
+
+    /// Creates a new `GraphemeU8` from a `char8`.
+    ///
+    /// # Errors
+    /// Returns [`MismatchedCapacity`] if `CAP` > 255
+    /// or < `c.`[`len_utf8()`][char8#method.len_utf8].
+    ///
+    /// Will always succeed if `CAP` >= 2 and <= 255.
+    pub const fn from_char8(c: char8) -> Result<Self, MismatchedCapacity> {
+        Ok(Self(unwrap![ok? StringU8::from_char8(c)]))
+    }
+
+    /// Creates a new `GraphemeU8` from a `char16`.
+    ///
+    /// # Errors
+    /// Returns [`MismatchedCapacity`] if `CAP` > 255
+    /// or < `c.`[`len_utf8()`][char16#method.len_utf8].
+    ///
+    /// Will always succeed if `CAP` >= 3 and <= 255.
+    pub const fn from_char16(c: char16) -> Result<Self, MismatchedCapacity> {
+        Ok(Self(unwrap![ok? StringU8::from_char16(c)]))
+    }
+
+    /// Creates a new `GraphemeU8` from a `char`.
+    /// # Errors
+    /// Returns [`MismatchedCapacity`] if `CAP` > 255
+    /// or < `c.`[`len_utf8()`][UnicodeScalar#method.len_utf8].
+    ///
+    /// Will never panic if `CAP` >= 4 and <= 255.
+    pub const fn from_char(c: char) -> Result<Self, MismatchedCapacity> {
+        Ok(Self(unwrap![ok? StringU8::from_char(c)]))
+    }
 
     //